% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcond.R
\name{fitcond}
\alias{fitcond}
\title{Fits cuticular and intercellular conductance to data
\code{fitcond} Fits cuticular and intercellular conductances}
\usage{
fitcond(data, rdiff, gsw, gcw, gias)
}
\arguments{
\item{data}{A dataframe.}

\item{rdiff}{Difference in total and stomatal resistances to CO2 diffusion}

\item{gsw}{Stomatal conductance to water}

\item{gcw}{Cuticular conductance to water}

\item{gias}{Intercellular conductance to water?}
}
\value{
fitcond fits cuticular and intercellular conductances
}
\description{
Fits cuticular and intercellular conductance to data
\code{fitcond} Fits cuticular and intercellular conductances
}
