% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcond2.R
\name{fitcond2}
\alias{fitcond2}
\title{Fits cuticular and intercellular conductance to data
\code{fitcond2} Fits cuticular and intercellular conductances}
\usage{
fitcond2(data, rtc..rtc, gsw, gcw, gias, a, b, c)
}
\arguments{
\item{data}{A dataframe.}

\item{rtc..rtc}{Difference in total and stomatal resistances to CO2 diffusion}

\item{gsw}{Stomatal conductance to water}

\item{gcw}{Cuticular conductance to water}

\item{gias}{Intercellular conductance to water?}

\item{a}{Initial value for gcw in curve fitting}

\item{b}{Initial value for gias in curve fitting}

\item{c}{exponential factor accounting for pathlength complexity in gias}
}
\value{
fitcond fits cuticular and intercellular conductances
}
\description{
Fits cuticular and intercellular conductance to data
\code{fitcond2} Fits cuticular and intercellular conductances
}
