% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcond3.R
\name{fitcond3}
\alias{fitcond3}
\title{Fits cuticular and intercellular conductance to data
\code{fitcond3} Fits cuticular and intercellular conductances}
\usage{
fitcond3(data, rtc..rtc, gsw, gcw, gias, a, b)
}
\arguments{
\item{data}{A dataframe.}

\item{rtc..rtc}{Difference in total and stomatal resistances to CO2 diffusion}

\item{gsw}{Stomatal conductance to water}

\item{gcw}{Cuticular conductance to water}

\item{gias}{Intercellular conductance to water?}

\item{a}{Initial value for gcw in curve fitting}

\item{b}{Initial value for gias in curve fitting}
}
\value{
fitcond3 fits cuticular and intercellular conductances
}
\description{
Fits cuticular and intercellular conductance to data
\code{fitcond3} Fits cuticular and intercellular conductances
}
