% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcond4.R
\name{fitcond4}
\alias{fitcond4}
\title{Fits cuticular and intercellular conductance to data
\code{fitcond4} Fits cuticular and intercellular conductances}
\usage{
fitcond4(data, rtc..rtc, gsw, gcw, x, a, b)
}
\arguments{
\item{data}{A dataframe.}

\item{rtc..rtc}{Difference in total and stomatal resistances to CO2 diffusion}

\item{gsw}{Stomatal conductance to water}

\item{gcw}{Cuticular conductance to water}

\item{x}{Intercellular conductance to water?}

\item{a}{Initial value for gcw in curve fitting}

\item{b}{Initial value for gias in curve fitting}
}
\value{
fitcond4 fits cuticular and intercellular conductances
}
\description{
Fits cuticular and intercellular conductance to data
\code{fitcond4} Fits cuticular and intercellular conductances
}
