% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfit.R
\name{plotfit}
\alias{plotfit}
\title{Plots fitcond
\code{plotfit} Plots fitcond}
\usage{
plotfit(data, fit)
}
\arguments{
\item{data}{A dataframe.}

\item{fit}{The output from fitcond}
}
\value{
plotfit plots the output of fitcond
}
\description{
Plots fitcond
\code{plotfit} Plots fitcond
}
