% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procharley.R
\name{procharley}
\alias{procharley}
\title{Calculating mesophyll conductance via fluorescence
\code{procharley} Calculates mesophyll conductance and chloroplastic CO2}
\usage{
procharley(data, a, b, gstar, gea, gc, Rd, Tleaf)
}
\arguments{
\item{data}{Dataframe}

\item{a}{leaf absoptance}

\item{b}{fraction of absorbed quanta that reaches PSII}

\item{gstar}{CO2 compensation point in absence of Rd}

\item{gea}{activation energy for gstar}

\item{gc}{constant for the temperature response of gstar}

\item{Rd}{dark respiration}

\item{Tleaf}{specifies leaf temperature to use}
}
\description{
Calculating mesophyll conductance via fluorescence
\code{procharley} Calculates mesophyll conductance and chloroplastic CO2
}
